feat: add methods description in IDocxTemplate interface
Added XML comments with brief descriptions to all public methods in the IDocxTemplate interface. The purpose of these changes is to provide developers with a hand on understanding on what individual methods in the interface do, what parameters they require, and what these parameters represent. Descriptions of the methods take into account the purpose of the methods and the impact they make on the document template.

Asana: https://app.asana.com/0/1203851531040615/1205133534920134/f