feat: add Exist(string key) method across configurations
In the JsonOneFileConfiguration, ZipJsonConfiguration, InMemoryConfiguration, and JsonConfiguration classes, the Exist<TPocoType>(string key) method has been overloaded to Exist(string key). The main aim was to abstract the specifics of type-intensive operations to a more generalized method. The IConfiguration interface was also altered to reflect the change and the older method was marked as [Obsolete]. ValidateKey is also called at the beginnings of the overloaded Exist methods to ensure key integrity.