feat(cfg): add optional sorting to JSON configuration file
An optional sort function was added to the JsonOneFileConfiguration constructor. A boolean parameter "sortKeysInFile" was added to directly control whether the keys should be sorted before they are written into the file. This was done to provide more flexibility in handling JSON files as in some use-cases, a sorted JSON is easier to read and debug.