fix: add AngleMs error message to resources
Added a new error message, `AngleMs_ErrorMessage`, to both the Russian and English resources. This was done to improve the user's understanding of the correct input format when they provide an angle in Minute Second (MS) format. Updated the `GetErrorMessage` function in AngleMs.cs to return the new message.

Asana: https://app.asana.com/0/1203851531040615/1205715713006945/f